renamed postcard mini to linkPreview
