+++ src/components/index.tsx
@@ -112,7 +112,7 @@ import OrDivider from './orDivider/orDividerView';
 import PostTranslationModal from './post-translation-modal/postTranslationModal';
 import { ImageViewer } from './imageViewer';
 import { WalkthroughMarker } from './walkthroughMarker';
-import { PostCardMini } from './postCardMini';
+import { LinkPreview } from './linkPreview';
 
 // Basic UI Elements
 import {
@@ -277,5 +277,5 @@ export {
   WalkthroughMarker,
   ProposalVoteRequest,
   HiveAuthModal,
-  PostCardMini,
+  LinkPreview,
 };
