+++ src/components/linkPreview/index.ts
@@ -1,0 +1,1 @@
+export * from './container/linkPreview';
