--- src/components/postCardMini/index.ts
@@ -1,1 +1,0 @@
-export * from './container/postCardMini';
